Moving more tests around.
