+++ node/libs/roles/src/validator/messages/testonly.rs
@@ -1,40 +1,168 @@
-use super::*;
-use crate::validator;
-use crate::validator::messages::v1::{testonly::validator_committee, LeaderSelectionMode};
-use zksync_consensus_crypto::Text;
+use super::{
+    v1, Block, BlockNumber, ChainId, Committee, ForkNumber, Genesis, GenesisHash, GenesisRaw,
+    Justification, Msg, MsgHash, NetAddress, Payload, PayloadHash, PreGenesisBlock,
+    ProtocolVersion, Signed, WeightedValidator,
+};
+use crate::validator::SecretKey;
+use rand::{
+    distributions::{Distribution, Standard},
+    Rng,
+};
+use zksync_concurrency::time;
+use zksync_consensus_utils::enum_util::Variant;
 
-/// Hardcoded payload.
-pub(crate) fn payload() -> Payload {
-    Payload(
-        hex::decode("57b79660558f18d56b5196053f64007030a1cb7eeadb5c32d816b9439f77edf5f6bd9d")
-            .unwrap(),
-    )
-}
-
-/// Hardcoded validator secret keys.
-pub(crate) fn validator_keys() -> Vec<validator::SecretKey> {
-    [
-        "validator:secret:bls12_381:27cb45b1670a1ae8d376a85821d51c7f91ebc6e32788027a84758441aaf0a987",
-        "validator:secret:bls12_381:20132edc08a529e927f155e710ae7295a2a0d249f1b1f37726894d1d0d8f0d81",
-        "validator:secret:bls12_381:0946901f0a6650284726763b12de5da0f06df0016c8ec2144cf6b1903f1979a6",
-        "validator:secret:bls12_381:3143a64c079b2f50545288d7c9b282281e05c97ac043228830a9660ddd63fea3",
-        "validator:secret:bls12_381:5512f40d33844c1c8107aa630af764005ab6e13f6bf8edb59b4ca3683727e619",
-    ]
-    .iter()
-    .map(|raw| Text::new(raw).decode().unwrap())
-    .collect()
-}
-
-/// Hardcoded genesis with no attesters.
-pub(crate) fn genesis() -> Genesis {
-    GenesisRaw {
-        chain_id: ChainId(1337),
-        fork_number: ForkNumber(42),
-        first_block: BlockNumber(2834),
-
-        protocol_version: ProtocolVersion(1),
-        validators: validator_committee(),
-        leader_selection: LeaderSelectionMode::Weighted,
+impl Distribution<BlockNumber> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> BlockNumber {
+        BlockNumber(rng.gen())
     }
-    .with_hash()
+}
+
+impl Distribution<ProtocolVersion> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ProtocolVersion {
+        ProtocolVersion(rng.gen())
+    }
+}
+
+impl Distribution<ForkNumber> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ForkNumber {
+        ForkNumber(rng.gen())
+    }
+}
+
+impl Distribution<ChainId> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ChainId {
+        ChainId(rng.gen())
+    }
+}
+
+impl Distribution<GenesisHash> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> GenesisHash {
+        GenesisHash(rng.gen())
+    }
+}
+
+impl Distribution<GenesisRaw> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> GenesisRaw {
+        let mut genesis = GenesisRaw {
+            chain_id: rng.gen(),
+            fork_number: rng.gen(),
+            first_block: rng.gen(),
+
+            protocol_version: rng.gen(),
+            validators: rng.gen(),
+            leader_selection: rng.gen(),
+        };
+
+        // In order for the genesis to be valid, sticky/rota leaders need to be in the validator committee.
+        if let v1::LeaderSelectionMode::Sticky(_) = genesis.leader_selection {
+            let i = rng.gen_range(0..genesis.validators.len());
+            genesis.leader_selection =
+                v1::LeaderSelectionMode::Sticky(genesis.validators.get(i).unwrap().key.clone());
+        } else if let v1::LeaderSelectionMode::Rota(pks) = genesis.leader_selection {
+            let n = pks.len();
+            let mut pks = Vec::new();
+            for _ in 0..n {
+                let i = rng.gen_range(0..genesis.validators.len());
+                pks.push(genesis.validators.get(i).unwrap().key.clone());
+            }
+            genesis.leader_selection = v1::LeaderSelectionMode::Rota(pks);
+        }
+
+        genesis
+    }
+}
+
+impl Distribution<Genesis> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Genesis {
+        rng.gen::<GenesisRaw>().with_hash()
+    }
+}
+
+impl Distribution<PayloadHash> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> PayloadHash {
+        PayloadHash(rng.gen())
+    }
+}
+
+impl Distribution<Payload> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Payload {
+        let size: usize = rng.gen_range(500..1000);
+        Payload((0..size).map(|_| rng.gen()).collect())
+    }
+}
+
+impl Distribution<Justification> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Justification {
+        let size: usize = rng.gen_range(500..1000);
+        Justification((0..size).map(|_| rng.gen()).collect())
+    }
+}
+
+impl Distribution<PreGenesisBlock> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> PreGenesisBlock {
+        PreGenesisBlock {
+            number: rng.gen(),
+            payload: rng.gen(),
+            justification: rng.gen(),
+        }
+    }
+}
+
+impl Distribution<Block> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Block {
+        match rng.gen_range(0..2) {
+            0 => Block::PreGenesis(rng.gen()),
+            _ => Block::FinalV1(rng.gen()),
+        }
+    }
+}
+
+impl Distribution<Committee> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Committee {
+        let count = rng.gen_range(1..11);
+        let public_keys = (0..count).map(|_| WeightedValidator {
+            key: rng.gen(),
+            weight: 1,
+        });
+        Committee::new(public_keys).unwrap()
+    }
+}
+
+impl Distribution<NetAddress> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> NetAddress {
+        NetAddress {
+            addr: std::net::SocketAddr::new(
+                std::net::IpAddr::from(rng.gen::<[u8; 16]>()),
+                rng.gen(),
+            ),
+            version: rng.gen(),
+            timestamp: time::UNIX_EPOCH + time::Duration::seconds(rng.gen_range(0..1000000000)),
+        }
+    }
+}
+
+impl Distribution<Msg> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Msg {
+        match rng.gen_range(0..3) {
+            0 => Msg::ConsensusV1(rng.gen()),
+            1 => Msg::SessionId(rng.gen()),
+            2 => Msg::NetAddress(rng.gen()),
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl Distribution<MsgHash> for Standard {
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> MsgHash {
+        MsgHash(rng.gen())
+    }
+}
+
+impl<V: Variant<Msg>> Distribution<Signed<V>> for Standard
+where
+    Standard: Distribution<V>,
+{
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signed<V> {
+        rng.gen::<SecretKey>().sign_msg(rng.gen())
+    }
 }
