+++ node/libs/roles/src/validator/testonly.rs
@@ -1,23 +1,14 @@
 //! Test-only utilities.
-use super::messages::v1::{
-    BlockHeader, CommitQC, ConsensusMsg, FinalBlock, LeaderProposal, LeaderSelectionMode, Phase,
-    ProposalJustification, ReplicaCommit, ReplicaNewView, ReplicaTimeout, Signers, TimeoutQC, View,
-    ViewNumber,
+use super::v1::{
+    BlockHeader, CommitQC, FinalBlock, LeaderSelectionMode, ReplicaCommit, ReplicaTimeout,
+    TimeoutQC, View, ViewNumber,
 };
 use super::{
-    AggregateSignature, Block, BlockNumber, ChainId, Committee, ForkNumber, Genesis, GenesisHash,
-    GenesisRaw, Justification, Msg, MsgHash, NetAddress, Payload, PayloadHash, PreGenesisBlock,
-    ProofOfPossession, ProtocolVersion, PublicKey, SecretKey, Signature, Signed, WeightedValidator,
+    Block, BlockNumber, ChainId, Committee, ForkNumber, Genesis, GenesisRaw, Payload,
+    PreGenesisBlock, ProtocolVersion, SecretKey, WeightedValidator,
 };
 use crate::attester;
-use bit_vec::BitVec;
-use rand::{
-    distributions::{Distribution, Standard},
-    Rng,
-};
-use std::sync::Arc;
-use zksync_concurrency::time;
-use zksync_consensus_utils::enum_util::Variant;
+use rand::Rng;
 
 /// Test setup specification.
 #[derive(Debug, Clone)]
@@ -63,6 +54,26 @@ impl SetupSpec {
             protocol_version: ProtocolVersion::CURRENT,
             leader_selection: LeaderSelectionMode::RoundRobin,
         }
+    }
+}
+
+/// Setup.
+#[derive(Debug, Clone)]
+pub struct SetupInner {
+    /// Validators' secret keys.
+    pub validator_keys: Vec<SecretKey>,
+    /// Attesters' secret keys.
+    pub attester_keys: Vec<attester::SecretKey>,
+    /// Past blocks.
+    pub blocks: Vec<Block>,
+    /// Genesis config.
+    pub genesis: Genesis,
+}
+
+impl std::ops::Deref for Setup {
+    type Target = SetupInner;
+    fn deref(&self) -> &Self::Target {
+        &self.0
     }
 }
 
@@ -284,368 +295,5 @@ impl Setup {
         }
 
         qc
-    }
-}
-
-/// Setup.
-#[derive(Debug, Clone)]
-pub struct SetupInner {
-    /// Validators' secret keys.
-    pub validator_keys: Vec<SecretKey>,
-    /// Attesters' secret keys.
-    pub attester_keys: Vec<attester::SecretKey>,
-    /// Past blocks.
-    pub blocks: Vec<Block>,
-    /// Genesis config.
-    pub genesis: Genesis,
-}
-
-impl std::ops::Deref for Setup {
-    type Target = SetupInner;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl AggregateSignature {
-    /// Generate a new aggregate signature from a list of signatures.
-    pub fn aggregate<'a>(sigs: impl IntoIterator<Item = &'a Signature>) -> Self {
-        let mut agg = Self::default();
-        for sig in sigs {
-            agg.add(sig);
-        }
-        agg
-    }
-}
-
-impl Distribution<AggregateSignature> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> AggregateSignature {
-        AggregateSignature(rng.gen())
-    }
-}
-
-impl Distribution<PublicKey> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> PublicKey {
-        PublicKey(rng.gen())
-    }
-}
-
-impl Distribution<Signature> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signature {
-        Signature(rng.gen())
-    }
-}
-
-impl Distribution<ProofOfPossession> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ProofOfPossession {
-        ProofOfPossession(rng.gen())
-    }
-}
-
-impl Distribution<SecretKey> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> SecretKey {
-        SecretKey(Arc::new(rng.gen()))
-    }
-}
-
-impl Distribution<BlockNumber> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> BlockNumber {
-        BlockNumber(rng.gen())
-    }
-}
-
-impl Distribution<ProtocolVersion> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ProtocolVersion {
-        ProtocolVersion(rng.gen())
-    }
-}
-
-impl Distribution<ForkNumber> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ForkNumber {
-        ForkNumber(rng.gen())
-    }
-}
-
-impl Distribution<ChainId> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ChainId {
-        ChainId(rng.gen())
-    }
-}
-
-impl Distribution<GenesisHash> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> GenesisHash {
-        GenesisHash(rng.gen())
-    }
-}
-
-impl Distribution<LeaderSelectionMode> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> LeaderSelectionMode {
-        match rng.gen_range(0..=3) {
-            0 => LeaderSelectionMode::RoundRobin,
-            1 => LeaderSelectionMode::Sticky(rng.gen()),
-            3 => LeaderSelectionMode::Rota({
-                let n = rng.gen_range(1..=3);
-                rng.sample_iter(Standard).take(n).collect()
-            }),
-            _ => LeaderSelectionMode::Weighted,
-        }
-    }
-}
-
-impl Distribution<GenesisRaw> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> GenesisRaw {
-        let mut genesis = GenesisRaw {
-            chain_id: rng.gen(),
-            fork_number: rng.gen(),
-            first_block: rng.gen(),
-
-            protocol_version: rng.gen(),
-            validators: rng.gen(),
-            leader_selection: rng.gen(),
-        };
-
-        // In order for the genesis to be valid, sticky/rota leaders need to be in the validator committee.
-        if let LeaderSelectionMode::Sticky(_) = genesis.leader_selection {
-            let i = rng.gen_range(0..genesis.validators.len());
-            genesis.leader_selection =
-                LeaderSelectionMode::Sticky(genesis.validators.get(i).unwrap().key.clone());
-        } else if let LeaderSelectionMode::Rota(pks) = genesis.leader_selection {
-            let n = pks.len();
-            let mut pks = Vec::new();
-            for _ in 0..n {
-                let i = rng.gen_range(0..genesis.validators.len());
-                pks.push(genesis.validators.get(i).unwrap().key.clone());
-            }
-            genesis.leader_selection = LeaderSelectionMode::Rota(pks);
-        }
-
-        genesis
-    }
-}
-
-impl Distribution<Genesis> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Genesis {
-        rng.gen::<GenesisRaw>().with_hash()
-    }
-}
-
-impl Distribution<PayloadHash> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> PayloadHash {
-        PayloadHash(rng.gen())
-    }
-}
-
-impl Distribution<BlockHeader> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> BlockHeader {
-        BlockHeader {
-            number: rng.gen(),
-            payload: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<Payload> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Payload {
-        let size: usize = rng.gen_range(500..1000);
-        Payload((0..size).map(|_| rng.gen()).collect())
-    }
-}
-
-impl Distribution<FinalBlock> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> FinalBlock {
-        FinalBlock {
-            payload: rng.gen(),
-            justification: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<Justification> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Justification {
-        let size: usize = rng.gen_range(500..1000);
-        Justification((0..size).map(|_| rng.gen()).collect())
-    }
-}
-
-impl Distribution<PreGenesisBlock> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> PreGenesisBlock {
-        PreGenesisBlock {
-            number: rng.gen(),
-            payload: rng.gen(),
-            justification: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<Block> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Block {
-        match rng.gen_range(0..2) {
-            0 => Block::PreGenesis(rng.gen()),
-            _ => Block::FinalV1(rng.gen()),
-        }
-    }
-}
-
-impl Distribution<ReplicaTimeout> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ReplicaTimeout {
-        ReplicaTimeout {
-            view: rng.gen(),
-            high_vote: rng.gen(),
-            high_qc: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<ReplicaCommit> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ReplicaCommit {
-        ReplicaCommit {
-            view: rng.gen(),
-            proposal: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<ReplicaNewView> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ReplicaNewView {
-        ReplicaNewView {
-            justification: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<LeaderProposal> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> LeaderProposal {
-        LeaderProposal {
-            proposal_payload: rng.gen(),
-            justification: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<TimeoutQC> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> TimeoutQC {
-        let n = rng.gen_range(1..11);
-        let map = (0..n).map(|_| (rng.gen(), rng.gen())).collect();
-
-        TimeoutQC {
-            view: rng.gen(),
-            map,
-            signature: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<CommitQC> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> CommitQC {
-        CommitQC {
-            message: rng.gen(),
-            signers: rng.gen(),
-            signature: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<ProposalJustification> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ProposalJustification {
-        match rng.gen_range(0..2) {
-            0 => ProposalJustification::Commit(rng.gen()),
-            1 => ProposalJustification::Timeout(rng.gen()),
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl Distribution<Signers> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signers {
-        Signers(BitVec::from_bytes(&rng.gen::<[u8; 4]>()))
-    }
-}
-
-impl Distribution<Committee> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Committee {
-        let count = rng.gen_range(1..11);
-        let public_keys = (0..count).map(|_| WeightedValidator {
-            key: rng.gen(),
-            weight: 1,
-        });
-        Committee::new(public_keys).unwrap()
-    }
-}
-
-impl Distribution<ViewNumber> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ViewNumber {
-        ViewNumber(rng.gen())
-    }
-}
-
-impl Distribution<View> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> View {
-        View {
-            genesis: rng.gen(),
-            number: rng.gen(),
-        }
-    }
-}
-
-impl Distribution<Phase> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Phase {
-        let i = rng.gen_range(0..2);
-
-        match i {
-            0 => Phase::Prepare,
-            1 => Phase::Commit,
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl Distribution<NetAddress> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> NetAddress {
-        NetAddress {
-            addr: std::net::SocketAddr::new(
-                std::net::IpAddr::from(rng.gen::<[u8; 16]>()),
-                rng.gen(),
-            ),
-            version: rng.gen(),
-            timestamp: time::UNIX_EPOCH + time::Duration::seconds(rng.gen_range(0..1000000000)),
-        }
-    }
-}
-
-impl Distribution<Msg> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Msg {
-        match rng.gen_range(0..3) {
-            0 => Msg::ConsensusV1(rng.gen()),
-            1 => Msg::SessionId(rng.gen()),
-            2 => Msg::NetAddress(rng.gen()),
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl Distribution<ConsensusMsg> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> ConsensusMsg {
-        match rng.gen_range(0..4) {
-            0 => ConsensusMsg::LeaderProposal(rng.gen()),
-            1 => ConsensusMsg::ReplicaCommit(rng.gen()),
-            2 => ConsensusMsg::ReplicaNewView(rng.gen()),
-            3 => ConsensusMsg::ReplicaTimeout(rng.gen()),
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl Distribution<MsgHash> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> MsgHash {
-        MsgHash(rng.gen())
-    }
-}
-
-impl<V: Variant<Msg>> Distribution<Signed<V>> for Standard
-where
-    Standard: Distribution<V>,
-{
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signed<V> {
-        rng.gen::<SecretKey>().sign_msg(rng.gen())
     }
 }
